refactor(hikcamera-sys): regenerate status codes as i32 + sys crate cleanup
The HikCamera MVS SDK declares status-code constants as unsigned
0x80000000-style literals, but its functions return int (signed). Bindgen
cannot emit those literals as i32 -- Rust rejects out-of-range integer
literals -- so previous versions of this crate generated them as u32 and
sprinkled 'as u32' casts across the safe wrapper.

build.rs now parses MvErrorDefine.h + MvISPErrorDefine.h itself and emits
each status code as 'pub const MV_E_HANDLE: i32 = 0x80000000_u32 as i32'.
bindgen's blocklist_file keeps it from double-emitting those constants.
The result: status codes are i32, matching the SDK's c_int return value,
with zero casts at call sites throughout the wrapper.

Also in this commit (sys-crate hygiene):
- Add .clangd so editors can resolve wrapper.h's MvCameraControl.h include
- Add wrapper.h #error guard for non-Windows targets (defense in depth)
- Add README.md (workspace readme field pointed at a missing file)
- Add tests/compile_pass.rs smoke tests for FFI symbols and i32 status codes
- Add status-code count drift detector (>=200 codes expected)
- Drop target-lexicon build-dependency; use CARGO_CFG_TARGET_*
- Drop the CargoCallbacks layer; rerun-if-changed covers the same ground
- Generate review copies under target/generated/ (stable path for diffing)
- Rewrite crate-level docs explaining the bindings/status-codes split
- .gitignore: add .zed/ (local editor config), drop stale bindings.rs rule

Wrapper changes:
- Status(pub i32), check(code: i32), HikCameraError::code() -> Option<i32>
- SDK_ERRORS lookup table typed [(i32, StatusInfo)]

Site docs:
- New developer/bindings-generation page (EN + ZH) covering the design
  rationale, the three bindgen alternatives that were rejected, and the
  traps the split avoids.
- Architecture page updated; lib.rs is no longer a one-liner.
- Developer index links to the new page.

+++ crates/hikcamera/src/error_tests.rs
@@ -2,31 +2,31 @@ use super::*;
 
 #[test]
 fn ok_code() {
-    assert!(check(sys::MV_OK as i32).is_ok());
+    assert!(check(sys::MV_OK).is_ok());
 }
 
 #[test]
 fn status_ok() {
     assert!(Status::OK.is_ok());
     assert!(Status(0).is_ok());
-    assert!(!Status(sys::MV_E_NODATA as i32).is_ok());
+    assert!(!Status(sys::MV_E_NODATA).is_ok());
 }
 
 #[test]
 fn status_display() {
-    let s = Status(sys::MV_E_LOAD_LIBRARY as i32);
+    let s = Status(sys::MV_E_LOAD_LIBRARY);
     assert_eq!(s.to_string(), "MV_E_LOAD_LIBRARY (-2147483636, 0x8000000C)");
 }
 
 #[test]
 fn sdk_error() {
-    let status = Status(sys::MV_E_NODATA as i32);
+    let status = Status(sys::MV_E_NODATA);
     let error = HikCameraError::Sdk { status };
 
     assert_eq!(status.info().name, "MV_E_NODATA");
     assert_eq!(status.info().message, "Timeout or no data received");
-    assert_eq!(error.code(), Some(sys::MV_E_NODATA as i32));
-    assert_eq!(check(sys::MV_E_NODATA as i32), Err(error));
+    assert_eq!(error.code(), Some(sys::MV_E_NODATA));
+    assert_eq!(check(sys::MV_E_NODATA), Err(error));
 }
 
 #[test]
@@ -42,7 +42,7 @@ fn unknown_sdk_error() {
 #[test]
 fn display_sdk_error() {
     let error = HikCameraError::Sdk {
-        status: Status(sys::MV_E_LOAD_LIBRARY as i32),
+        status: Status(sys::MV_E_LOAD_LIBRARY),
     };
 
     assert_eq!(
